made config creation synchronous to avoid a race condition
