+++ ags/app.ts
@@ -1,32 +1,19 @@
 import { App } from "astal/gtk3"
 import Bar from "@Bar/Bar"
 import NotificationCenter from "@Notif/NotificationCenter"
-import { exec, readFileAsync, Variable } from "astal"
+import { exec, Variable } from "astal"
 import PowerMenu from "@Power/PowerMenu"
 import ControlCenter from "@CC/ControlCenter"
-import { config, theme, themeOpts, wallpaper, wallpaperOpts } from "./variables/theme-variables"
 import AppLauncher from "@/AppLauncher/AppLauncher"
 import WallpaperSwitcher from "@/WallpaperSwitcher/WallpaperSwitcher"
 import NotificationPopups from "@/Notification/Popups"
+import { readConfig, readCurrentTheme } from "./helpers/init"
 
 exec(["sass", "./style.scss", "/tmp/style.css"])
 
 // config creation
-readFileAsync(`${SRC}/currentTheme.json`).then(data => {
-    const parsed: { name: string, wallpaper: string } = JSON.parse(data)
-    theme.set(parsed.name)
-    wallpaper.set(parsed.wallpaper)
-})
-
-readFileAsync(`${SRC}/themes.json`).then(data => {
-    const parsed: Array<Config> = JSON.parse(data)
-    config.set(parsed)
-
-    themeOpts.set(parsed.map(t => t.name))
-    wallpaperOpts.set(parsed.filter(t => t.name === theme.get())[0].wallpapers)
-})
-
-
+readCurrentTheme()
+readConfig()
 
 const menuState = Variable<string>("none")
 
@@ -58,13 +45,13 @@ App.start({
         })
         App.get_monitors().map(monitor => {
             ControlCenter({ gdkmonitor: monitor, menuState: menuState })
-        }),
+        })
         App.get_monitors().map(monitor => {
             AppLauncher({ gdkmonitor: monitor, menuState: menuState })
-        }),
+        })
         App.get_monitors().map(monitor => {
             WallpaperSwitcher({ gdkmonitor: monitor, menuState: menuState })
-        }),
+        })
         NotificationPopups(App.get_monitors()[0])
     },
 })
