+++ hypr/hyprland.conf
@@ -50,7 +50,7 @@ $menu = wofi --show drun
 
 exec-once = swww-daemon
 # exec-once = agsv1
-exec-once = ags
+exec-once = ags run
 exec-once = hypridle
 exec-once = dbus-update-activation-environment --systemd WAYLAND_DISPLAY XDG_CURRENT_DESKTOP
 # exec-once = nm-applet &
